fix libnoise performance bug

+++ ManicDiggerLib/LibNoise/IModule.cs
@@ -2,5 +2,5 @@ namespace LibNoise;
 
 public interface IModule
 {
-	double GetValue(double x, double y, double z);
+	float GetValue(float x, float y, float z);
 }
